wsd: jail: move linkOrCopy to JailUtil
